# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bookmarks - Social Ebook Reader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  description: string;
  coverColor: string;
  quoteCount: number;
  commentCount: number;
  highlightCount: number;
}

export interface Quote {
  id: number;
  canonicalBookId: number;
  text: string;
  /** Distinctive leading substring used to locate the quote in an EPUB */
  searchText: string;
  highlightCount: number;
  commentCount: number;
  /** 0.0 to 1.0 representing community highlight strength */
  highlightIntensity: number;
  highlightedByMe?: boolean;
}

export interface Comment {
  id: number;
  quoteId: number;
  userId: number;
  username: string;
  avatarColor: string;
  text: string;
  quoteText?: string;
  likeCount: number;
  likedByMe: boolean;
  savedByMe: boolean;
  createdAt: string;
}

export interface BookDetail {
  id: number;
  title: string;
  author: string;
  description: string;
  coverColor: string;
  quoteCount: number;
  commentCount: number;
  highlightCount: number;
  topQuotes: Quote[];
  bestComments: Comment[];
}

export interface CommentInput {
  /** @minLength 1 */
  text: string;
}

export interface CreateQuoteInput {
  /**
     * @minLength 8
     * @maxLength 1000
     */
  text: string;
}

export interface HighlightInput {
  userLibraryId?: number;
  cfiRange?: string;
}

export interface HighlightResult {
  highlighted: boolean;
  highlightCount: number;
}

export interface CreateBookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  description?: string;
  isbn?: string;
}

export interface CreateLibraryInput {
  /** The presigned upload URL (or object path) returned after upload */
  uploadURL: string;
}

export interface ManualMatchInput {
  canonicalBookId: number;
}

export interface LibraryEntry {
  id: number;
  userId: number;
  /** @nullable */
  canonicalBookId: number | null;
  epubObjectPath: string;
  epubUrl: string;
  /** @nullable */
  originalTitle: string | null;
  /** @nullable */
  originalAuthor: string | null;
  /** @nullable */
  originalIsbn: string | null;
  /** @nullable */
  lastReadingLocation: string | null;
  createdAt: string;
  book: Book | null;
}

export interface UpdateReadingLocationInput {
  /** @minLength 1 */
  location: string;
}

export type MatchResultStatus = typeof MatchResultStatus[keyof typeof MatchResultStatus];


export const MatchResultStatus = {
  matched: 'matched',
  candidates: 'candidates',
  none: 'none',
} as const;

export interface MatchResult {
  status: MatchResultStatus;
  book: Book | null;
  candidates: Book[];
}

export interface LibraryUploadResult {
  entry: LibraryEntry;
  match: MatchResult;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}

export interface SaveResult {
  saved: boolean;
}

export interface UserRef {
  userId: number;
}

export interface User {
  id: number;
  username: string;
  avatarColor: string;
  createdAt: string;
  /** HMAC-signed bearer token — send as Authorization header */
  token: string;
}

export interface UserWithFriendStatus {
  id: number;
  username: string;
  avatarColor: string;
  createdAt: string;
  isFriend: boolean;
}

export interface UserInput {
  /** @minLength 2 */
  username: string;
}

export interface FriendInput {
  friendId: number;
}

export type ListBooksParams = {
q?: string;
};

export type GetQuoteCommentsParams = {
filter?: GetQuoteCommentsFilter;
};

export type GetQuoteCommentsFilter = typeof GetQuoteCommentsFilter[keyof typeof GetQuoteCommentsFilter];


export const GetQuoteCommentsFilter = {
  best: 'best',
  friends: 'friends',
  all: 'all',
} as const;

export type SearchUsersParams = {
q: string;
userId?: number;
};