# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bookmarks - Social Ebook Reader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List / search canonical books
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a canonical book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookInput"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{bookId}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book with its best comments and most-highlighted quotes
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "404":
          description: Book not found
...
  /quotes/{quoteId}/comments:
    get:
      operationId: getQuoteComments
      tags: [comments]
      summary: Get comments for a quote
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [best, friends, all]
            default: all
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment on a quote
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /comments/{commentId}/like:
    post:
      operationId: likeComment
      tags: [comments]
      summary: Like a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /comments/{commentId}/save:
    post:
      operationId: saveComment
      tags: [comments]
      summary: Save/unsave a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
...
components:
  schemas:
...
    Book:
      type: object
      required: [id, title, author, description, coverColor, quoteCount, commentCount, highlightCount]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        description:
          type: string
        coverColor:
          type: string
        quoteCount:
          type: integer
        commentCount:
          type: integer
        highlightCount:
          type: integer
    Quote:
      type: object
      required: [id, canonicalBookId, text, searchText, highlightCount, commentCount, highlightIntensity]
      properties:
        id:
          type: integer
        canonicalBookId:
          type: integer
        text:
          type: string
        searchText:
          type: string
          description: Distinctive leading substring used to locate the quote in an EPUB
        highlightCount:
          type: integer
        commentCount:
          type: integer
        highlightIntensity:
          type: number
          description: 0.0 to 1.0 representing community highlight strength
        highlightedByMe:
          type: boolean
    BookDetail:
      type: object
      required: [id, title, author, description, coverColor, quoteCount, commentCount, highlightCount, topQuotes, bestComments]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        description:
          type: string
        coverColor:
          type: string
        quoteCount:
          type: integer
        commentCount:
          type: integer
        highlightCount:
          type: integer
        topQuotes:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        bestComments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    Comment:
      type: object
      required: [id, quoteId, userId, username, avatarColor, text, likeCount, likedByMe, savedByMe, createdAt]
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarColor:
          type: string
        text:
          type: string
        quoteText:
          type: string
        likeCount:
          type: integer
        likedByMe:
          type: boolean
        savedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
...
    CreateBookInput:
      type: object
      required: [title, author]
      properties:
        title:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1
        description:
          type: string
        isbn:
          type: string
...
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
    SaveResult:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
...