# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bookmarks - Social Ebook Reader API
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /** @minLength 1 */
  text: string;
}